+++ AssemblyLine/Patches/InventoryGuiPatch.cs
@@ -1,94 +1,83 @@
 ﻿namespace AssemblyLine;
 
+using System.Collections.Generic;
 using System.Linq;
+using System.Reflection.Emit;
+
+using ComfyLib;
 
 using HarmonyLib;
 
 using UnityEngine;
-using UnityEngine.Events;
-using UnityEngine.UI;
 
 [HarmonyPatch(typeof(InventoryGui))]
 static class InventoryGuiPatch {
   [HarmonyPostfix]
   [HarmonyPatch(nameof(InventoryGui.Show))]
-  public static void ShowPostfix(InventoryGui __instance, Container container) {
-    if (CraftingManager.IncrementButtonTransform != null
-        || CraftingManager.DecrementButtonTransform != null
-        || CraftingManager.CraftButton != null) {
-      return;
-    }
-
-    CraftingManager.CraftButton = __instance.m_craftButton;
-
-    CraftingManager.CreateCountText();
-    CraftingManager.SetCraftAmountToMin();
-    CraftingManager.ScaleCraftButton(__instance.m_craftButton);
-
-    CraftingManager.IncrementButtonTransform = CraftingManager.CreateButton(__instance, "incrementButton", "+");
-    CraftingManager.DecrementButtonTransform = CraftingManager.CreateButton(__instance, "decrementButton", "-");
-
-    CraftingManager.IncrementButton = CraftingManager.IncrementButtonTransform.GetComponent<Button>();
-    CraftingManager.DecrementButton = CraftingManager.DecrementButtonTransform.GetComponent<Button>();
-
-    CraftingManager.IncrementButtonTransform.pivot = new Vector2(1f, 0.75f);
-    CraftingManager.DecrementButtonTransform.pivot = new Vector2(1f, 0f);
-
-    CraftingManager.IncrementButtonTransform.GetComponent<Button>()
-        .onClick.AddListener(new UnityAction(CraftingManager.OnIncrementPressed));
-    CraftingManager.DecrementButtonTransform.GetComponent<Button>()
-        .onClick.AddListener(new UnityAction(CraftingManager.OnDecrementPressed));
+  static void ShowPostfix(InventoryGui __instance, Container container) {
+    CraftingManager.CreateOrSetupUI(__instance);
   }
 
   [HarmonyPostfix]
   [HarmonyPatch(nameof(InventoryGui.OnTabCraftPressed))]
-  public static void OnTabCraftPressedPostfix(InventoryGui __instance) {
+  static void OnTabCraftPressedPostfix(InventoryGui __instance) {
     CraftingManager.SetButtonInteractable(true);
   }
 
   [HarmonyPostfix]
   [HarmonyPatch(nameof(InventoryGui.OnTabUpgradePressed))]
-  public static void OnTabUpgradePressedPostfix(InventoryGui __instance) {
+  static void OnTabUpgradePressedPostfix(InventoryGui __instance) {
     CraftingManager.SetButtonInteractable(false);
     CraftingManager.SetCraftAmountToMin();
   }
 
   [HarmonyPostfix]
   [HarmonyPatch(nameof(InventoryGui.OnCraftPressed))]
-  public static void OnCraftPressedPrefix(InventoryGui __instance) {
+  static void OnCraftPressedPostfix(InventoryGui __instance) {
     CraftingManager.CraftsRemaining = CraftingManager.CountValue - 1;
+  }
+
+  [HarmonyTranspiler]
+  [HarmonyPatch(nameof(InventoryGui.OnCraftPressed))]
+  static IEnumerable<CodeInstruction> OnCraftPressedTranspiler(IEnumerable<CodeInstruction> instructions) {
+    return new CodeMatcher(instructions)
+        .Start()
+        .MatchStartForward(
+            new CodeMatch(OpCodes.Ldc_I4_0),
+            new CodeMatch(
+                OpCodes.Stfld,
+                AccessTools.Field(typeof(InventoryGui), nameof(InventoryGui.m_multiCrafting))),
+            new CodeMatch(OpCodes.Ldarg_0))
+        .ThrowIfInvalid($"Could not patch InventoryGui.OnCraftPressed()! (get-button-alt-place)")
+        .Advance(offset: 2)
+        .InsertAndAdvance(
+            new CodeInstruction(OpCodes.Ldarg_0),
+            new CodeInstruction(
+                OpCodes.Call,
+                AccessTools.Method(typeof(InventoryGuiPatch), nameof(SetIsMultiCraftingDelegate))),
+            new CodeInstruction(
+                OpCodes.Stfld,
+                AccessTools.Field(typeof(InventoryGui), nameof(InventoryGui.m_multiCrafting))))
+        .InstructionEnumeration();
   }
 
   [HarmonyPostfix]
   [HarmonyPatch(nameof(InventoryGui.OnCraftCancelPressed))]
-  public static void OnCraftCancelPressedPostfix(InventoryGui __instance) {
+  static void OnCraftCancelPressedPostfix(InventoryGui __instance) {
     CraftingManager.CraftsRemaining = 0;
   }
 
   [HarmonyPostfix]
   [HarmonyPatch(nameof(InventoryGui.Hide))]
-  public static void HidePostfix(InventoryGui __instance) {
+  static void HidePostfix(InventoryGui __instance) {
     CraftingManager.CraftsRemaining = 0;
   }
 
   [HarmonyPostfix]
   [HarmonyPatch(nameof(InventoryGui.UpdateRecipe))]
-  public static void UpdateRecipePostfix(InventoryGui __instance, Player player, float dt) {
-    if (__instance == null
-        || Player.m_localPlayer == null
-        || Player.m_localPlayer != player
-        || Player.m_localPlayer.GetInventory() == null
-        || __instance.m_selectedRecipe.Recipe == null) {
-
+  static void UpdateRecipePostfix(InventoryGui __instance, Player player, float dt) {
+    if (!__instance.m_selectedRecipe.Recipe) {
       return;
-    }
-
-    if (__instance.InCraftTab()
-        && CraftingManager.IncrementButton != null
-        && CraftingManager.DecrementButton != null) {
-      CraftingManager.
-              IncrementButton.interactable = true;
-      CraftingManager.DecrementButton.interactable = true;
     }
 
     if (!player.GetInventory().CanAddItem(
@@ -102,18 +91,46 @@ static class InventoryGuiPatch {
     if (__instance.m_craftTimer < 0 && CraftingManager.CraftsRemaining > 0) {
       CraftingManager.CraftsRemaining--;
       __instance.m_craftTimer = 0;
+
       CraftingManager.DecrementCraftAmount(1);
       CraftingManager.UpdateMaxAmount();
       CraftingManager.SetRequirementText();
 
-      if (Player.m_localPlayer.GetCurrentCraftingStation() != null) {
-        player.GetCurrentCraftingStation().m_craftItemEffects.Create(
-            player.transform.position, Quaternion.identity, null, 1f, -1);
-      }
+      CraftingStation craftingStation = player.GetCurrentCraftingStation();
 
-      __instance.m_craftItemEffects.Create(player.transform.position, Quaternion.identity, null, 1f, -1);
+      EffectList craftItemEffects =
+          craftingStation ? craftingStation.m_craftItemEffects : __instance.m_craftItemEffects;
+
+      craftItemEffects?.Create(player.transform.position, Quaternion.identity);
+
       __instance.m_craftRecipe = __instance.m_selectedRecipe.Recipe;
     }
+  }
+
+  [HarmonyTranspiler]
+  [HarmonyPatch(nameof(InventoryGui.UpdateRecipe))]
+  static IEnumerable<CodeInstruction> UpdateRecipeTranspiler(IEnumerable<CodeInstruction> instructions) {
+    return new CodeMatcher(instructions)
+        .Start()
+        .MatchStartForward(
+            new CodeMatch(OpCodes.Ldc_I4_0),
+            new CodeMatch(OpCodes.Stloc_S),
+            new CodeMatch(OpCodes.Ldloc_2),
+            new CodeMatch(OpCodes.Brtrue))
+        .ThrowIfInvalid($"Could not patch InventoryGui.UpdateRecipe()! (get-button-alt-place)")
+        .Advance(offset: 1)
+        .SaveInstruction(out CodeInstruction stLocSv5Instruction)
+        .Advance(offset: 1)
+        .InsertAndAdvance(
+            new CodeInstruction(
+                OpCodes.Call,
+                AccessTools.Method(typeof(InventoryGuiPatch), nameof(SetIsMultiCraftingDelegate))),
+            stLocSv5Instruction)
+        .InstructionEnumeration();
+  }
+
+  static bool SetIsMultiCraftingDelegate() {
+    return false;
   }
 
   [HarmonyPostfix]
@@ -135,18 +152,19 @@ static class InventoryGuiPatch {
 
   [HarmonyPostfix]
   [HarmonyPatch(nameof(InventoryGui.SetupRequirement))]
-  public static void SetupRequirementPostfix(
+  static void SetupRequirementPostfix(
       InventoryGui __instance, Transform elementRoot, Piece.Requirement req, Player player, bool craft, int quality) {
     if (!InventoryGui.IsVisible()) {
       return;
     }
 
-    if (!CraftingManager.RequirementAmountByName.Keys.Contains(req.m_resItem.m_itemData.m_shared.m_name)) {
-      CraftingManager.RequirementAmountByName.Add(req.m_resItem.m_itemData.m_shared.m_name, req.GetAmount(quality));
+    string resourceItemName = req.m_resItem.m_itemData.m_shared.m_name;
+
+    if (!CraftingManager.RequirementAmountByName.Keys.Contains(resourceItemName)) {
+      CraftingManager.RequirementAmountByName.Add(resourceItemName, req.GetAmount(quality));
       CraftingManager.MaxAmountByName.Add(
-          req.m_resItem.m_itemData.m_shared.m_name,
-          Player.m_localPlayer.GetInventory().CountItems(req.m_resItem.m_itemData.m_shared.m_name));
-      CraftingManager.RequirementTransformByName.Add(req.m_resItem.m_itemData.m_shared.m_name, elementRoot);
+          resourceItemName, Player.m_localPlayer.GetInventory().CountItems(resourceItemName));
+      CraftingManager.RequirementTransformByName.Add(resourceItemName, elementRoot);
     }
 
     CraftingManager.SetRequirementText();